LPS-50491 - Move logger files to portal/util/logger
