+++ portal-web/test/functional/com/liferay/portalweb/util/logger/LoggerHandler.java
@@ -12,7 +12,9 @@
  * details.
  */
 
-package com.liferay.portalweb.portal.util.liferayselenium;
+package com.liferay.portalweb.util.logger;
+
+import com.liferay.portalweb.portal.util.liferayselenium.LiferaySelenium;
 
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.InvocationTargetException;
